+++ packages/keyring-internal-api/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Uncategorized
+
+- chore: add `yarn constraints` support ([#514](https://github.com/MetaMask/accounts/pull/514))
+
 ### Added
 
 - Add internal Stellar account struct (`InternalXlmAccount`) ([#483](https://github.com/MetaMask/accounts/pull/483))